requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.24.0
httpx>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.1
